created adapter, model, service & ui packages
