+++ app/src/main/java/com/staff/staffapp/ui/MainActivity.java
@@ -1,4 +1,4 @@
-package com.staff.staffapp;
+package com.staff.staffapp.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,7 +7,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.staff.staffapp.FAQ.FAQ;
+import com.staff.staffapp.R;
 
 public class MainActivity extends AppCompatActivity {
     private Button mFAQButton;
