+++ app/src/main/res/drawable/rectanglebackgroundwhite.xml
@@ -2,4 +2,8 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <solid android:color="@color/white"/>
 
+    <corners
+        android:topLeftRadius="30dp"
+        android:topRightRadius="30dp"/>
+
 </shape>
