+++ app/src/main/res/layout/activity_ambulance.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FAQ.Ambulance">
+    tools:context=".ui.Ambulance">
 
     <com.github.barteksc.pdfviewer.PDFView
         android:id="@+id/pdfAmbulance"
