+++ app/src/main/res/layout/activity_contacts.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FAQ.Contacts"
+    tools:context=".ui.Contacts"
     android:gravity="center"
     android:layout_above="@id/bottomNav">
