+++ app/src/main/res/layout/activity_eastern.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FAQ.Eastern">
+    tools:context=".ui.Eastern">
 
     <com.github.barteksc.pdfviewer.PDFView
         android:id="@+id/pdfEastern"
