+++ app/src/main/res/layout/activity_ethics.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FAQ.Ethics"
+    tools:context=".ui.Ethics"
     android:gravity="center"
     android:layout_above="@id/bottomNav">
