+++ app/src/main/res/layout/activity_jobs.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FAQ.Jobs"
+    tools:context=".ui.Jobs"
     android:gravity="center"
     android:layout_above="@id/bottomNav"
     android:background="#DADFE5">
@@ -15,7 +15,7 @@
         android:layout_gravity="center"
         android:textColor="@color/colorBitterSweet"
         android:textStyle="bold"
-        android:background="@drawable/jobserch"
+        android:background="@drawable/hiring1"
         android:textSize="30sp"/>
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
