+++ app/src/main/res/layout/activity_leave.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FAQ.Leave"
+    tools:context=".ui.Leave"
     android:gravity="center"
     android:layout_above="@id/bottomNav">
