+++ app/src/main/res/layout/activity_panel.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FAQ.Panel">
+    tools:context=".ui.Panel">
 
     <com.github.barteksc.pdfviewer.PDFView
         android:id="@+id/pdfAttachment1"
