+++ app/src/main/res/layout/activity_pharmacies.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FAQ.Pharmacies">
+    tools:context=".ui.Pharmacies">
 
     <com.github.barteksc.pdfviewer.PDFView
         android:id="@+id/pdfPharmacies"
